accordion typescript conversion
