--- src/components/Accordion/Accordion.reducers.js
@@ -1,6 +1,0 @@
-export function onSelect(state = {}, selectedIndex) {
-	return {
-		...state,
-		selectedIndex,
-	};
-}
