+++ src/components/Accordion/Accordion.reducers.ts
@@ -1,0 +1,8 @@
+import { IAccordionState } from './Accordion';
+
+export function onSelect(state: IAccordionState, selectedIndex: number) {
+	return {
+		...state,
+		selectedIndex,
+	};
+}
